Refactored publisher index routine.
